requirements.txt
django
djangorestframework
psycopg2-binary
celery
redis
openpyxl
django-celery-results